M3-392 - fix: rename page obj, update imports, removed hardcoded assertion
* rename page obj, update imports, removed hard-coded string assertion
